requirements.txt
blinker==1.7.0
click==8.1.7
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
fsspec==2023.12.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
packaging==23.2
python-dotenv==1.0.0
regex==2023.12.25
sympy==1.12
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
